Changed so much lets just check this in.
svn path=/src/trunk/; revision=17595

+++ tools/mozilla/build_ff3.1_langs.sh
@@ -23,10 +23,17 @@
 # http://translate.sourceforge.net/wiki/toolkit/mozilla_l10n_scripts     #
 ##########################################################################
 
-BUILD_DIR="/path/to/build_root"
-MOZCENTRAL_DIR="${BUILD_DIR}/mozilla-central" # Change "../mozilla-central" on line 39 too if you change this var
-#HG_LANGS="af ar as be bg bn-IN ca cs da de el en-GB en-ZA es-AR es-ES et eu fa fi fr fy-NL ga-IE gl gu-IN he hi-IN hu hy-AM id is it ja ja-JP-mac ka kn ko ku langs lt lv mk ml mn mr nb-NO ne-NP nl nn-NO nr nso pa-IN pl ro ru rw si sk sl sq sr ss st sv-SE ta te th tn tr ts uk ve xh zh-CN zh-TW zu"
-HG_LANGS="af"
+. ~/dev/lm/trunk/setpath
+
+if [ $# -eq 0 ]; then
+	#HG_LANGS="af ak am en-ZA ff km lg nso ur son st-LS su sw zu"
+	HG_LANGS="af ak am en-ZA ff lg nso ur son st-LS su sw wo zu"
+else
+	HG_LANGS=$*
+fi
+
+BUILD_DIR="/home/dwayne/dev/mozilla"
+MOZCENTRAL_DIR="${BUILD_DIR}/mozilla-aurora" # Change "../mozilla-central" on line 39 too if you change this var
 L10N_DIR="${BUILD_DIR}/l10n"
 PO_DIR="${BUILD_DIR}/po"
 POPACK_DIR="${BUILD_DIR}/popacks"
@@ -34,9 +41,9 @@ PORECOVER_DIR="${BUILD_DIR}/po-recover"
 POT_INCLUDES="../README.mozilla-pot"
 POTPACK_DIR="${BUILD_DIR}/potpacks"
 POUPDATED_DIR="${BUILD_DIR}/po-updated"
-PRODUCT_DIRS="browser other-licenses/branding/firefox" # Directories in language repositories to clear before running po2moz
+PRODUCT_DIRS="browser dom netwerk other-licenses/branding/firefox security services/sync toolkit" # Directories in language repositories to clear before running po2moz
 LANGPACK_DIR="${BUILD_DIR}/xpi"
-FF_VERSION="3.5b4"
+FF_VERSION="4.0b7pre"
 
 # Include current dir in path (for buildxpi and others)
 CURDIR=`dirname $0`
@@ -68,23 +75,23 @@ do
 	(find ${lang} -name '*.orig' | xargs rm) || /bin/true
 done
 
-rm en-US
-rm -rf pot
+[ -d pot ] && rm -rf pot
 
 # en-US and all languages should be up-to-date now
 [ -d en-US_browser ] && rm -rf en-US_browser
-get_moz_enUS.py -s ../mozilla-central -d . -p browser -v
+[ -L en-US ] && rm en-US
+get_moz_enUS.py -s ../mozilla-aurora -d . -p browser -v
 mv en-US{,_browser}
 ln -sf en-US_browser ./en-US
 # CREATE POT FILES FROM en-US
-moz2po --progress=none -P --duplicates=msgctxt --exclude '.hg' en-US pot
+moz2po --errorlevel=traceback --progress=none -P --duplicates=msgctxt --exclude '.hg' en-US pot
 find pot \( -name '*.html.pot' -o -name '*.xhtml.pot' \) -exec rm -f {} \;
 
 # Create POT pack
 # Comment out the lines starting with "tar" and/or "zip" to keep from building archives in the specific format(s).
 PACKNAME="${POTPACK_DIR}/firefox-${FF_VERSION}-`date +%Y%m%d`"
-tar chjf ${PACKNAME}.tar.bz2 pot en-US ${POT_INCLUDES}
-zip -qr9 ${PACKNAME}.zip pot en-US ${POT_INCLUDES}
+#tar chjf ${PACKNAME}.tar.bz2 pot en-US ${POT_INCLUDES}
+#zip -qr9 ${PACKNAME}.zip pot en-US ${POT_INCLUDES}
 
 # The following functions are used in the loop following it
 function copyfile {
@@ -121,31 +128,62 @@ function copydir {
 
 for lang in ${HG_LANGS}
 do
-	## RECOVER - Recover PO files from existing l10n directory.
-	## Comment out the following "moz2po"-line if recovery should not be done.
-	[ ! -d ${PORECOVER_DIR}/${lang} ] && mkdir -p ${PORECOVER_DIR}/${lang}
-	moz2po --progress=none --errorlevel=traceback --duplicates=msgctxt --exclude=".#*" --exclude='.hg' \
-		-t ${L10N_DIR}/en-US ${L10N_DIR}/${lang} ${PORECOVER_DIR}/${lang}
+	### RECOVER - Recover PO files from existing l10n directory.
+	### Comment out the following "moz2po"-line if recovery should not be done.
+	#[ ! -d ${PORECOVER_DIR}/${lang} ] && mkdir -p ${PORECOVER_DIR}/${lang}
+	#moz2po --progress=none --errorlevel=traceback --duplicates=msgctxt --exclude=".#*" --exclude='.hg' \
+	#	-t ${L10N_DIR}/en-US ${L10N_DIR}/${lang} ${PORECOVER_DIR}/${lang}
 
-	[ ! -d ${PO_DIR}/${lang} ] && cp -R ${PORECOVER_DIR}/${lang} ${PO_DIR}
+	#[ ! -d ${PO_DIR}/${lang} ] && cp -R ${PORECOVER_DIR}/${lang} ${PO_DIR}
 
 	# Try and update existing PO files
+        polang=$(echo $lang|sed "s/-/_/g")
 	updated=""
-	[ -z ${updated} ] && [ -d ${PO_DIR}/${lang}/CVS ] && (cd ${PO_DIR}/${lang}; cvs up) && updated="1"
-	[ -z ${updated} ] && [ -d ${PO_DIR}/${lang}/.hg ] && (cd ${PO_DIR}/${lang}; hg pull -u) && updated="1"
-	[ -z ${updated} ] && [ -d ${PO_DIR}/${lang}/.svn ] && (cd ${PO_DIR}/${lang}; svn up) && updated="1"
+	[ -z ${updated} ] && [ -d ${PO_DIR}/${polang}/CVS ] && (cd ${PO_DIR}/${polang}; cvs up) && updated="1"
+	[ -z ${updated} ] && [ -d ${PO_DIR}/${polang}/.hg ] && (cd ${PO_DIR}/${polang}; hg pull -u) && updated="1"
+	[ -z ${updated} ] && [ -d ${PO_DIR}/${polang}/.svn ] && (cd ${PO_DIR}/${polang}; svn up) && updated="1"
 
 	# Copy directory structure while preserving version control metadata
-	rm -rf ${POUPDATED_DIR}/${lang}
-	cp -R ${PO_DIR}/${lang} ${POUPDATED_DIR}
-	find ${POUPDATED_DIR}/${lang} -name '*.po' -exec rm -f {} \;
+	rm -rf ${POUPDATED_DIR}/${polang}
+	cp -R ${PO_DIR}/${polang} ${POUPDATED_DIR}
+	find ${POUPDATED_DIR}/${polang} -name '*.po' -exec rm -f {} \;
 
 	## MIGRATE - Migrate PO files to new POT files.
 	# Comment out the following "pomigrate2"-line if migration should not be done.
 	tempdir=`mktemp -d`
-	cp -R ${PO_DIR}/${lang} ${tempdir}/${lang}
-	pomigrate2 --use-compendium --quiet --pot2po ${tempdir}/${lang} ${POUPDATED_DIR}/${lang} ${L10N_DIR}/pot
+	cp -R ${PO_DIR}/${polang} ${tempdir}/${polang}
+	pomigrate2 --use-compendium --pot2po --quiet ${tempdir}/${polang} ${POUPDATED_DIR}/${polang} ${L10N_DIR}/pot
 	rm -rf ${tempdir}
+
+	## Cleanup migrated PO files
+	# msgcat to make them look the same
+	for po in $(find ${POUPDATED_DIR}/${polang} -name "*.po")
+	do
+		msgcat $po > $po.2
+		mv $po.2 $po
+	done
+	# Revert files with only header changes
+	svn revert $(svn diff --diff-cmd diff -x "--unified=3 --ignore-matching-lines=POT-Creation --ignore-matching-lines=X-Generator -s" ${POUPDATED_DIR}/${polang} |
+	egrep "are identical$" |
+	sed -r "s/^Files //;s/\s+\(revision.*$//")
+
+	## Migrate to new PO files: move old to obsolete/ and add new files
+	(cd ${POUPDATED_DIR}/${polang}
+	svn revert -R mail editor other-licenses/branding/thunderbird
+	for newfile in $(svn status . | egrep "^\?" | egrep "\.po$" | sed "s/\?\w*//")
+	do
+		svn add $newfile
+	done
+
+	svn revert -R obsolete
+	mkdir -p obsolete
+	svn add obsolete
+	for oldfile in $(svn status . | egrep "^!"| egrep "\.po$" | sed "s/!\w*//")
+	do
+		svn revert $oldfile
+		svn move --parents $oldfile obsolete/$oldfile
+	done
+	)
 
 	# Pre-po2moz hacks
 	lang_product_dirs=
@@ -153,10 +191,10 @@ do
 	[ -d ${L10N_DIR}/${lang} ] && find ${lang_product_dirs} \( -name '*.dtd' -o -name '*.properties' \) -exec rm -f {} \;
 	find ${POUPDATED_DIR} \( -name '*.html.po' -o -name '*.xhtml.po' \) -exec rm -f {} \;
 
-	## PO2MOZ - Create Mozilla l10n layout from migrated PO files.
+	# PO2MOZ - Create Mozilla l10n layout from migrated PO files.
 	# Comment out the "po2moz"-line below to prevent l10n files to be updated to the current PO files.
-	#po2moz --progress=none --errorlevel=traceback --exclude=".svn" --exclude=".hg" \
-	#	-t ${L10N_DIR}/en-US -i ${POUPDATED_DIR}/${lang} -o ${L10N_DIR}/${lang}
+	po2moz --progress=none --errorlevel=traceback --exclude=".svn" --exclude=".hg" --exclude="obsolete" --exclude="editor" --exclude="mail" --exclude="thunderbird" \
+		-t ${L10N_DIR}/en-US -i ${POUPDATED_DIR}/${polang} -o ${L10N_DIR}/${lang}
 
 	# Copy files not handled by moz2po/po2moz
 	copydir browser/os2 ${lang}
@@ -167,22 +205,26 @@ do
 	copyfile browser/microsummary-generators/list.txt ${lang}
 	copyfile browser/profile/chrome/userChrome-example.css ${lang}
 	copyfile browser/profile/chrome/userContent-example.css ${lang}
-	copyfile browser/searchplugins/list.txt ${lang}
-	copyfile extensions/reporter/chrome/reporterOverlay.properties ${lang}
+        # Ignore lists.txt since we need specil approval for that
+	#copyfile browser/searchplugins/list.txt ${lang}
 	copyfile toolkit/chrome/global/intl.css ${lang}
-	[ ! -f ${L10N_DIR}/${lang}/browser/profile/bookmarks.html ] && copyfile browser/profile/bookmarks.html ${lang}
-	sed -i "s/en-US/${lang}/g" ${L10N_DIR}/${lang}/browser/profile/bookmarks.html || /bin/true
-	
+        # Revert some files that need careful human review or authorisation
+	[ -d ${L10N_DIR}/${lang}/.hg ] && (cd ${L10N_DIR}/${lang}; hg revert browser/chrome/browser-region/region.properties browser/searchplugins/list.txt)
+	# These seem to have been removed for Fx4
+	#[ ! -f ${L10N_DIR}/${lang}/browser/profile/bookmarks.html ] && copyfile browser/profile/bookmarks.html ${lang}
+	#sed -i "s/en-US/${lang}/g" ${L10N_DIR}/${lang}/browser/profile/bookmarks.html || /bin/true
+
 	## CREATE PO PACK - Create archives of PO files.
 	# Comment out the lines starting with "tar" and/or "zip" to keep from building archives in the specific format(s).
-	PACKNAME="${POPACK_DIR}/firefox-${FF_VERSION}-${lang}-`date +%Y%m%d`"
+	PACKNAME="${POPACK_DIR}/firefox-${FF_VERSION}-${polang}-`date +%Y%m%d`"
 	(
 		cd ${BUILD_DIR}
-		tar cjf ${PACKNAME}.tar.bz2 --exclude '.svn' --exclude '.hg' ${L10N_DIR_REL}/${lang} ${POUPDATED_DIR_REL}/${lang}
-		zip -qr9 ${PACKNAME}.zip ${L10N_DIR_REL}/${lang} ${POUPDATED_DIR_REL}/${lang} -x '*.svn*' -x "*.hg*"
+		#tar cjf ${PACKNAME}.tar.bz2 --exclude '.svn' --exclude '.hg' ${L10N_DIR_REL}/${lang} ${POUPDATED_DIR_REL}/${polang}
+		#zip -qr9 ${PACKNAME}.zip ${L10N_DIR_REL}/${lang} ${POUPDATED_DIR_REL}/${polang} -x '*.svn*' -x "*.hg*"
 	)
 
 	## CREATE XPI LANGPACK
 	# Comment out the "buildxpi"-line below if XPI langpacks should not be built.
 	#buildxpi.py -L ${L10N_DIR} -s ${MOZCENTRAL_DIR} -o ${LANGPACK_DIR} ${lang}
+
 done
